requirements.txt
alembic==1.0.2
amqp==2.4.1
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.5
beautifulsoup4==4.7.1
billiard==3.5.0.5
blinker==1.4
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
factory-boy==2.11.1
Faker==0.9.2
Flask==1.0.2
Flask-API==1.0
Flask-Cors==3.0.6
Flask-JWT-Extended==3.15.0
Flask-Migrate==2.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.8.1
idna==2.8
inflect==1.0.1
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.3.0
mailchimp3==3.0.7
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==4.3.0
paystackapi==1.2.7
pluggy==0.8.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.7.0
pycparser==2.19
PyJWT==1.6.4
PySocks==1.6.8
pytest==3.10.0
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.9
redis==3.2.0
requests==2.21.0
six==1.11.0
soupsieve==1.9
SQLAlchemy==1.2.13
stripe==2.21.0
termcolor==1.1.0
text-unidecode==1.2
twilio==6.25.0
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
